Move all route definition and logic to a separate class
